Disallow slurm.conf node configurations with NodeName=ALL.
The check is case insensitive and logs an error and propagates it upstream,
effectively resulting in a fatal when parsed from slurmctld or slurmd.
If not parsed from these daemons, the node entry is silently ignored.

The purpose is to make 'ALL' (and its case combinations) be a reserved word
in preparation for subsequent changes.

Bug 13099